Now writing original basename to source
Now writing original (from camera) basename to source since the field I was using "Title" is used for something else.

Don't know how to access source (and most other IPTC and EXIF) in photos easily, but it can be done by looking at the imported file.

Still working on script to convert existing Aperture to something that Photos will like. That is moving Title to Source and Caption to Title.